+++ src/lib/ThemeContext.tsx
@@ -1,13 +1,7 @@
-import {isWeb} from 'platform/detection'
 import React, {ReactNode, createContext, useContext} from 'react'
-import {
-  AppState,
-  TextStyle,
-  useColorScheme as useColorScheme_BUGGY,
-  ViewStyle,
-  ColorSchemeName,
-} from 'react-native'
+import {TextStyle, ViewStyle, ColorSchemeName} from 'react-native'
 import {darkTheme, defaultTheme} from './themes'
+import {useColorScheme_FIXED} from '#/lib/hooks/useColorScheme_FIXED'
 
 export type ColorScheme = 'light' | 'dark'
 
@@ -95,32 +89,6 @@ export const useTheme = () => useContext(ThemeContext)
 
 function getTheme(theme: ColorSchemeName) {
   return theme === 'dark' ? darkTheme : defaultTheme
-}
-
-/**
- * With RN iOS, we can only "trust" the color scheme reported while the app is
- * active. This is a workaround until the bug is fixed upstream.
- *
- * @see https://github.com/bluesky-social/social-app/pull/1417#issuecomment-1719868504
- * @see https://github.com/facebook/react-native/pull/39439
- */
-function useColorScheme_FIXED() {
-  const colorScheme = useColorScheme_BUGGY()
-  const [currentColorScheme, setCurrentColorScheme] =
-    React.useState<ColorSchemeName>(colorScheme)
-
-  React.useEffect(() => {
-    // we don't need to be updating state on web
-    if (isWeb) return
-    const subscription = AppState.addEventListener('change', state => {
-      const isActive = state === 'active'
-      if (!isActive) return
-      setCurrentColorScheme(colorScheme)
-    })
-    return () => subscription.remove()
-  }, [colorScheme])
-
-  return isWeb ? colorScheme : currentColorScheme
 }
 
 export const ThemeProvider: React.FC<ThemeProviderProps> = ({
