+++ src/view/nova/components/Box.tsx
@@ -1,0 +1,3 @@
+import {View} from 'react-native'
+import {styled} from '#/view/nova/system'
+export const Box = styled(View)
