+++ src/view/nova/components/Button.tsx
@@ -1,0 +1,9 @@
+import {Pressable} from 'react-native'
+import {styled} from '#/view/nova/system'
+
+export const Button = styled(Pressable, {
+  px: 'l',
+  py: 's',
+  bg: 'l3',
+  radius: 40,
+})
