+++ src/view/nova/util/platform.ts
